refactor: Combine common code in base.
