+++ web/api/chat_thread_handler.py
@@ -6,8 +6,9 @@ from docq.domain import FeatureKey, OrganisationFeatureType
 from pydantic import ValidationError
 from tornado.web import HTTPError
 
+from web.api.base import BaseRequestHandler
 from web.api.models import ThreadModel, ThreadPostRequestModel, ThreadResponseModel
-from web.api.utils import BaseRequestHandler, authenticated
+from web.api.utils import authenticated
 from web.utils.streamlit_application import st_app
